Patch to fix disabling mpi4py in cdms2.dataset
For `cdscan` to work when MPI is disabled, we need to be able
to call `setNetcdfUseParallelFlag(0)`.  Before this patch, this
call attempted to raise an error but there was a typo in the
error name (`CDSMError` instead of `CDMSError`), causing added
confusion.  We only want to raise an error if the code is
attempting to enable parallel NetCDF usage but MPI is disabled.